Track amount of work of chain
Rely on block header difficulty getter for positive quantity linearly
representing work.

Keep track of work in chain at genesis and at header insertion.

Store in chain service the header decorated with work (record
`#chain_header{}`).  Do not enrich main record `#header{}` as the
work-related field is not used outside the chain.

Adapt the type of the cached top header in the chain service to the
type of the stored headers.

Also:

* Split initialization function `do_init` in smaller functions - for
  readability;

* Refactor `top_header/0` and `top_block/0` - for easing type check by
  Dialyzer;

* Specify types of data stored in chain service;

* Clarify key of database of top header.