This commit introduces an implementation of the EVM and an initial parser and compiler for the Ring language.
The VM is called the EEEVM (Experimental Erlang Etherium VM).
Also known as AEVM.
The VM can handle almost all EVM operations except for
the operations on the block chain.
The VM passes over 500 EVM tests.

The parser handles a subset of the Ocaml like language Reason.
This subset will be extended with block chain operations to
giv the Ring language.

The compiler only compiles the identity function so far and
is only to be seen as a start for compilation from the Ring
AST to the AEVM.