Use add_and_ping_peers() (corrected) at startup; implemented re-pings
The aec_sync module would ping preconfigured peers one at a time.
This, combined with a lack of re-pings, could lead to an incomplete
discovery. Now, all preconfigured peers are first added to the peers
database, then pings are initiated.

The aec_peers server now sets re-ping timers for peers, using a
backoff algorithm if the ping failed (starting from MIN, then
increasing the wait in a sort-of-fibonacci sequence, up to MAX ms).
The MIN and MAX can be configured using the env var
{aecore, ping_interval_limits, {Min, Max}}, where Max >= Min, and
both values are in milliseconds. Default: {3000, 60000}.

If the ping succeeds, a re-ping is scheduled in MAX ms.