Teach `aec_miner` to search PoW solution for one nonce at a time
Handling of multiple nonces in PoW modules:

* Is not needed after introduction of block candidate, handled in
  `aec_miner`;

* Complicates the code.

Hence remove handling of multiple nonces in PoW modules.

The drawback in doing so is that the SHA-256 PoW becomes slower,
because the attempt for each nonce needs to be orchestrated by the
`aec_miner` process.  As the SHA-256 PoW is not currently used, the
benefits of this change outweigh the drawbacks.