Work around intermittent cuckoo cycle failures
With 5 parallel threads, the cuckoo cycle miner fails to
find an existing solution around 10% of the time (on my
machine), causing the tests in `aec_pow_cuckoo_tests`
and `aec_mining_tests` to fail.

The ideal solution to the problem would be to fix the race
conditions in the C++ code, but until that happens the
work around is to call the miner repeatedly (up to 10 times)
until it finds a solution.