Do not consume unexpected messages during OS process execution
While running PoW generation or validation in a separate OS process,
the code read unexpected messages in addition to the ones expected
from the process. This in not a problem for PoW generation as it runs
from a separate erlang process but the miner performs PoW validation
in its main process. It happened sometimes that the cast messages
the miner is sending to itself were consumed by the receive block
parsing the validation output, confusing the miner state machine.