Implement Cuckoo PoW verification in Erlang
The verification is a simple path traversal in a graph, the graph is
defined by the output of SipHash, and the SipHash keys are defined by
the BLAKE2b hash of the Header(hash) and the Nonce.

BLAKE2b and SipHash implementations have been verified by running a
QuickCheck property against reference implementations in C. A couple of
EUnit tests are included here for completeness.