Add target that developer is meant to run when pulling repo
When the aecuckoo Makefile points to a different git commit, a
non-clean workind directory `apps/aecuckoo/c_src` causes failure in
compilation.  A transparent handling of this would probably be complex
(e.g. unnecessary network interactions) and error-prone.

Introduce target that developer is meant to run after pulling repo and
wanting to be sure of not confusing the tooling.