Temporary patch for fetch_next/update_chain_from_pool
Adding blocks to the chain is done in the aec_sync gen_server-loop in
the current design. This patch makes us only add up to 20 blocks in one
go - making the gen_server more responsive. It also increases the
timeout for fetch_next (which was timing out frequently due to the block
adding) to 30s.

NOTE: these are not permanent fixes the program flow should be improved
to properly fix this...