Pt 156410826 sync fixes (#926)
* aec_peer_connection should cleanup on non-normal termination

* Fix error matching in get_header_by_X

* Make aec_peer_connection_listener trap exits

It should really be turned into a gen_server, and the child processes
should go in a separate supervision tree. This makes it behave better in
the meantime...

* Temporary patch for fetch_next/update_chain_from_pool

Adding blocks to the chain is done in the aec_sync gen_server-loop in
the current design. This patch makes us only add up to 20 blocks in one
go - making the gen_server more responsive. It also increases the
timeout for fetch_next (which was timing out frequently due to the block
adding) to 30s.

NOTE: these are not permanent fixes the program flow should be improved
to properly fix this...

* Fix bad test