Fix intermittent test failure in system test
Test case `docker_keeps_data` fails when stopping node executing
`stop_node(standalone_node, infinity, Cfg)`.  The reason is that the
`infinity` timeout is not cascaded to the `bin/epoch stop` command
execution, that is then executed with default timeout of 5s.  (I
measured locally between 6s and 10s.)

Such timeout happens also in test cases cleanup, where timeout is
specified as 0, and that is harmless and wanted - for completing test
run quickly.