Avoid HTTP request overhead
We are bloating the internal Cowboy routing state with many copies of
the JSON validator. This adds a significant overhead to HTTP requests
mainly because copying a data structure between processes means we
loose sharing and has to handle lots of memory. This fix is temporary,
it will parse the JSON schema for each request, that should be done by a
proper service.