PT-160377912 cheap encode decode (#1581)
* Factor out logic for RETURN instruction

* Shorten addresses in contract test output

* Add aeso_data:from_heap/3

* Add debug function to print a heap

* Read return values from the heap directly instead of encoding them in bytecode

Currently we don't have the return types for a contract, so the contract writes a
typerep to the heap. Eventually the return types should be in the contract metadata
saving us the heap encoding.

* Refactor: move the save_store function

* Benchmark to measure effects of optimisations

* Skip encoding of updated state

* Load calldata into memory before starting the VM

This means that you don't pay a gas cost for CALLDATACOPY, but you still pay
for the memory. It isn't really more expensive for the node to keep the calldata on
the heap than in the VM state, so I don't think makes sense.

* Add representation of 'typerep' to type representations

* Get rid of bytecode decoding of state

To make this work we have the init function return a typerep of the state type
and store that separately in the store. Eventually we should have the state type
stored in some contract metadata.

* Get rid of decoding of contract call return values

* Get rid of encoding of contract call arguments

This is the final bytecode encoding/decoding so cleans up a lot of code in the compiler.

* Remember modifiers (stateful, private, etc) when parsing

* Don't generate entry points for private functions

* Update tests to changes in Sophia execution

* Reduce some code duplication in eeevm_state

* Put calldata pointer on the stack instead of in memory

This gets rid of the base address 32 for calldata, so we can
use base address 0 for everything (and saves one word of memory)

* Remove gap in typerep encoding tags

This makes it encode as the corresponding recursive datatype.

* Release notes