Ditch the off-chain tx env
The off-chain calls must be ancored to a block hash, so we can reason for
time, block hash, beneficiary and etc. This implies that both participants share
the same fork of the chain and both had synced up to at least this block. Bearing
this in mind, it makes no sense to have different transaction enviroment for the
off-chain and the on-chain as they both are based on the same block hash.

Force progress calls are always pinned to the top.