Teach sync system test node ping interval
... by injecting its value from the test.

This makes the sync system test cases `net_split_recovery` and
`net_split_mining_power` wait for 10s (twice the ping interval) rather
than 240s.  This does not change the ping interval used on the nodes
themselves - that is still 5s.

This may exhibit failures not visible before (as test suite is now
stricter), and may require future fine-tuning of the test suite.