Stabilize sync system test for unexpected forks (#1609)
* Teach sync test to try harder to prevent forks

Sync system test failed once
[here](https://github.com/aeternity/epoch/blob/253f7d5b5ee621e99ddaa75cbfd48a55f886232d/system_test/aest_sync_SUITE.erl#L515)
on [CI](https://circleci.com/gh/aeternity/epoch/33830).  Node 1 and 2
of the first cluster were out of sync at height 40, even if test mines
up to 41 for attempting to resolve forks.  Block 39 was the same.
Blocks 40 and 41 on the two nodes were distinct - mined at max 500 ms
from each other (at same height).

This change introduces an extra block for aiming at resolving forks
more thoroughly.

The extra block is introduced even before the checks in try-until
loops, as the time the try-loop is attempted for may not be sufficient
for both mining and syncing.

This change also reviews the timeout used in try-until loops, as it
was relying on a wrong vision of the ping interval (120s instead of
the real 5s).  The value of the ping interval is clarified by getting
its value from the node.  This makes the sync system test cases
`net_split_recovery` and `net_split_mining_power` wait for 10s (twice
the ping interval) rather than 240s.  This does not change the ping
interval used on the nodes themselves, that is still 5s.  This may
exhibit failures not visible before (as test suite is now stricter),
and may require future fine-tuning of the test suite.

* Fix sync start time in sync system tests

Symptom was try-until giving up very early.

Analogous of 7a848434 for peers suite.

* Stabilize sync test by resolving peer name often

Default backoff times are very long (`[5000, 15000, 30000, 60000,
120000, 300000, 600000]`) compared to `2 * ping interval` (= 10s)
waited for by test.