Collect system info of Docker host of system tests
Save among artefacts of system test jobs the `dmesg` of the CI (Linux)
host.  This host is the same kernel as the Docker containers running
the epoch nodes so getting the `dmesg` of the CI host may collect info
re anomalies that may have happened inside containers.

I did not test that e.g. causing an out-of-memory inside a Docker
container is visible from the `dmesg` of the CI host.

I did not investigate whether the response of Docker Engine HTTP API
`/containers/<ID>/logs` contains `dmesg` portions already.