Change mechanism for creating calldata to allow arbitrary Sophia types
The most general way is to set the function name to "" and give a contract
as the argument containing a function __call(). For instance to call a function
foo with arguments arg1 : t1, arg2 : t2, and arg3 : t3, one would write

contract SomeName =
  // ... type defs required for t1, t2, t3
  function foo : (t1, t2, t3) => res
  function __call() = foo(arg1, arg2, arg3)

This will get type checked and appropriate calldata will be generated. A shorthand
is available, if no typedefs are required, where the function name argument is set to
"foo : (t1, t2, t3) => res" and the argument is "arg1, arg2, arg3".