Add minimal tx fee check to micro block validation
This change adds minimal tx fee check to the caller of
`aec_conductor:add_synced_block/1` and `aec_conductor:post_block/1` which
may offload the conductor process. The check is also performed before a micro
block is added to the database by conductor (but now it's less likely there
would be any errors).