Align CUDA miner configu instructions to changing app version
This assumes only one version of the app on the filesystem though it
seems reasonable, and not worth the effort of making the doc more
complex potentially causing losing the point in the section.

I tried executing the following analogous command (with wildcard) on
macOS (downloading Linux package for having the CUDA miner, though
that is orthogonal) and the miner is found by shell/time.
```
$ time ./lib/aecuckooprebuilt-*/priv/cuda29 -r 5
-bash: ./lib/aecuckooprebuilt-0.1.0/priv/cuda29: cannot execute binary file

real    0m0.003s
user    0m0.001s
sys     0m0.002s
```