Prefer miner with lower memory usage
Symptom:
```
CircleCI was unable to execute the job to completion, because the CircleCI agent received a kill signal midway through the job.
This means that that either the agent got killed accidentally (through a killall command), or that the build container entrypoint (i.e. PID 1) allocated too much memory or terminated early.  Consider clearing entrypoint/command values and try again.
```