Upgrade Sophia - add ecverify_secp256k1 to AEVM_3 (#2345)
* Upgrade Sophia - add ecverify_secp256k1 to AEVM_3

This includes bringing in a Minerva standalone compiler, and while at it also
a Roma standalone compiler that has the same interface. Also move from aehtt_logic
to aect_test_utils for all compilation. Phasing out aect_sophia...

* Use digest directly no point in double hashing

* Use the latest aesophia/aebytecode

* Fix sophia bytecode test re protocol version set

Assert that protocol version in test is set - for readability and
robustness of test.  Also fix test that was not setting it for not
exported CT callback.

* Use proper OTP 21 container for static analysis (#2352)

* Fix problems with empty string encoding in CLI compiler

* Roll back bytes(N) from VM-level