Wrap network and timing sensitive CI operations in a retry loop
Often CI jobs fail due to small random network errors or host CPU
stalls. This change introduces a simple retry loop around operations
which are prone to such issues. If a job still fails after a retry the
author of a change can at least be confident that the issue wasn't a
one-off problem.