Handle out of order async workers in fork signalling result computation
The results map in aec_fork_signalling process has 2 types of values:
* result entry - this is a tuple {Result, BlockHash, BlockHeight} where
  BlockHash and BlockHeight are associated with a pointer that points to
  this result entry. Based on the BlockHeight it is possible to keep
  just one pointer (the one on the top of the chain) pointing to this
  result entry;
* pointer - block hash of a block at height HE - 1

The results map's key is a block hash. It can be either a block hash
associated with HE - 1 block - it keeps result entry value; or
a block hash associated with a pointer (blocks between HE - 1 and
H2, exclusive).