Add database write retry
If the database encounters a write problem during execution of a
transaction the operation will exit and subsequently take down the node.
This change introduces the option to retry transactions to handle short
temporary disk write issues. At worst the previous behaviour will occur
after all retries have failed, at best the problem doesn't persist and
the write is successful after a retry.

The number of retries can be configured in the user configuration.