Add database write retry (#2984)
If the database encounters a write problem during execution of a
transaction the operation will exit and subsequently take down the node.
This change introduces the option to retry transactions to handle short
temporary disk write issues. At worst the previous behaviour will occur
after all retries have failed, at best the problem doesn't persist and
the write is successful after a retry.

The number of retries can be configured in the user configuration.

After running a transaction aec_db now queries an in-memory error store
for references to runtime errors concerning the objects which should
have been written during the respective transaction. These errors will
be written to the store by the database backend when encountered.

The error store handling and garbage collection is performed
in a separate process which is part of aecore.