Use -1 when name_ttl is not specified on name_update events
Right now, there's no way to distinguish on the env event if the AENS
update was called using `Some(0)` (update the name to expire
immediately) or `None` (don't update the name TTL).

This implementation now returns name_ttl = 0 when it was supposed to
expire immediately and `name_ttl = -1` when the name_ttl is supposed
to remain untouched.